♻️ refactor: OfflinePage migration
