+++ src/pages/_app.tsx
@@ -1,12 +1,10 @@
 import "@/styles/global.css";
 
 import type { AppProps } from "next/app";
-import { DefaultSeo } from "next-seo";
-import SEO from "next-seo.config";
 import { ThemeProvider } from "next-themes";
 
-import { SearchProvider } from "@/components/organisms/SearchArea/SearchContext";
-import { GoogleAnalytics, usePageView } from "@/lib/google-analytics";
+import { SearchProvider } from "@/components/feature/SearchArea/SearchContext";
+import { GoogleAnalytics } from "@/lib/google-analytics";
 
 // FIXME: MSW is broken "ERR_UNSUPPORTED_DIR_IMPORT"
 // if (process.env.NEXT_PUBLIC_MSW_ENABLED === "true") {
@@ -18,11 +16,8 @@ import { GoogleAnalytics, usePageView } from "@/lib/google-analytics";
 // }
 
 const App = ({ Component, pageProps }: AppProps) => {
-  usePageView();
-
   return (
     <div>
-      <DefaultSeo {...SEO} />
       <GoogleAnalytics />
       <ThemeProvider attribute="class" enableSystem>
         <SearchProvider>
